refactor(modules): convert CJS to ESM notation
